Add 'include collide.js' to usage instructions
For the swipe to work properly, we need collide.js. Currently, README.md only says to include 'ionic.tdcards.js' and 'ionic.tdcards.css'. This could easily cause new users to run into errors, which would be avoidable by adding 'collide.js' to the list of things to include under the Usage section in the Readme.